Move Validation class into Util package
